Made day of year float temporarily.

+++ mysite/pond.py
@@ -250,9 +250,10 @@ class Pond(object):
         Get Day of Year
         Getter method.
         @return: the day of on which measurements occurred.
-        @rtype: int
+        @rtype: float
         '''
-        return int(self.__day_of_year)
+        #TODO: fix this. only made it float so I could fix a temporary issue with the test data.
+        return float(self.__day_of_year)
 
 
     def get_length_of_day(self):
